fixed another couple of glitches

+++ tests/io/test_converters.py
@@ -42,6 +42,8 @@ def test_converter_json(knowledge_base):
 
     for n in range(0, 5):
         iob_file = files[n]
+        if not os.path.exists():
+            continue
         logger.info("{} {}".format(iob_file, brat_dir))
         dc.load(iob_file, standoff_dir=brat_dir)
         out = dc.to_json(output_dir)
